Changes for getting ready to remove SFArrayBase class

+++ src/apps/ethslurp/ethslurp.h
@@ -41,6 +41,6 @@ private:
 
 //---------------------------------------------------------------------------------------------------
 extern bool establishFolders(CToml& toml);
-extern void findBlockRange(const string_q& contents, uint32_t& minBlock, uint32_t& maxBlock);
+extern void findBlockRange(const string_q& contents, size_t& minBlock, size_t& maxBlock);
 
 extern CFileExportContext outScreen;
