+++ src/libs/acctlib/accountwatch.h
@@ -57,13 +57,13 @@ public:
 
     DECLARE_NODE(CAccountWatch);
 
-    const CBaseNode *getObjectAt(const string_q& fieldName, uint32_t index) const override;
+    const CBaseNode *getObjectAt(const string_q& fieldName, size_t index) const override;
 
     // EXISTING_CODE
     CAccountWatch(const string_q& _addr, const string_q& _name, blknum_t fB, blknum_t lB, const string_q& _color)
-    : address(toLower(_addr)), name(_name), color(_color), firstBlock(fB), lastBlock(lB) {}
-    string_q displayName(bool expand, bool terse, uint32_t w1=20, uint32_t w2=8) const { return displayName(expand, true,terse,w1,w2); }
-    string_q displayName(bool expand, bool useColor, bool terse, uint32_t w1=20, uint32_t w2=8) const;
+        : address(toLower(_addr)), name(_name), color(_color), firstBlock(fB), lastBlock(lB) {}
+    string_q displayName(bool expand, bool terse, size_t w1=20, size_t w2=8) const { return displayName(expand, true,terse,w1,w2); }
+    string_q displayName(bool expand, bool useColor, bool terse, size_t w1=20, size_t w2=8) const;
     SFBloom bloom;
     bool inBlock;
     CBalanceHistoryArray balanceHistory;
