+++ src/libs/acctlib/treeroot.cpp
@@ -66,7 +66,7 @@ bool CTreeRoot::setValueByName(const string_q& fieldName, const string_q& fieldV
                 root = new CTreeNode;
                 if (root) {
                     char *p = cleanUpJson((char *)fieldValue.c_str());
-                    uint32_t nFields = 0;
+                    size_t nFields = 0;
                     root->parseJson(p, nFields);
                     return true;
                 }
@@ -136,7 +136,7 @@ void CTreeRoot::registerClass(void) {
     if (been_here) return;
     been_here = true;
 
-    uint32_t fieldNum = 1000;
+    size_t fieldNum = 1000;
     ADD_FIELD(CTreeRoot, "schema",  T_NUMBER, ++fieldNum);
     ADD_FIELD(CTreeRoot, "deleted", T_BOOL,  ++fieldNum);
     ADD_FIELD(CTreeRoot, "showing", T_BOOL,  ++fieldNum);
