+++ src/libs/etherlib/block.h
@@ -54,7 +54,7 @@ public:
 
     DECLARE_NODE(CBlock);
 
-    const CBaseNode *getObjectAt(const string_q& fieldName, uint32_t index) const override;
+    const CBaseNode *getObjectAt(const string_q& fieldName, size_t index) const override;
 
     // EXISTING_CODE
     bool forEveryAddress      (ADDRESSFUNC func, TRANSFUNC filt, void *data);
