+++ src/libs/etherlib/function.h
@@ -50,7 +50,7 @@ public:
 
     DECLARE_NODE(CFunction);
 
-    const CBaseNode *getObjectAt(const string_q& fieldName, uint32_t index) const override;
+    const CBaseNode *getObjectAt(const string_q& fieldName, size_t index) const override;
 
     // EXISTING_CODE
     bool hasAddrs;
