+++ src/libs/etherlib/node_curl.h
@@ -28,7 +28,7 @@ namespace qblocks {
         string_q         result;
         string_q         provider;
         bool             is_error;
-        uint32_t         theID;
+        size_t           theID;
 
         CCurlContext(void);
         string_q getCurlID(void);
