+++ src/libs/etherlib/traceresult.cpp
@@ -127,7 +127,7 @@ void CTraceResult::registerClass(void) {
     if (been_here) return;
     been_here = true;
 
-    uint32_t fieldNum = 1000;
+    size_t fieldNum = 1000;
     ADD_FIELD(CTraceResult, "schema",  T_NUMBER, ++fieldNum);
     ADD_FIELD(CTraceResult, "deleted", T_BOOL,  ++fieldNum);
     ADD_FIELD(CTraceResult, "showing", T_BOOL,  ++fieldNum);
