+++ src/libs/utillib/accountname.cpp
@@ -129,7 +129,7 @@ void CAccountName::registerClass(void) {
     if (been_here) return;
     been_here = true;
 
-    uint32_t fieldNum = 1000;
+    size_t fieldNum = 1000;
     ADD_FIELD(CAccountName, "schema",  T_NUMBER, ++fieldNum);
     ADD_FIELD(CAccountName, "deleted", T_BOOL,  ++fieldNum);
     ADD_FIELD(CAccountName, "showing", T_BOOL,  ++fieldNum);
