+++ src/tools/getTrace/getTrace.cpp
@@ -64,7 +64,7 @@ bool visitTransaction(CTransaction& trans, void *data) {
     getTraces(traces, trans.getValueByName("hash"));
 
     cout << "[";
-    for (uint32_t i = 0 ; i < traces.size() ; i++) {
+    for (size_t i = 0 ; i < traces.size() ; i++) {
         traces[i].doExport(cout);
         cout << (i < traces.size()-1 ? ",\n" : "\n");
     }
