+++ src/tools/isContract/options.cpp
@@ -21,7 +21,7 @@ CParams params[] = {
     CParams("",              "Returns 'true' or 'false' if the given address(es) holds byte code "
                              "(optionally displays the code).\n"),
 };
-uint32_t nParams = sizeof(params) / sizeof(CParams);
+size_t nParams = sizeof(params) / sizeof(CParams);
 
 //---------------------------------------------------------------------------------------------------
 bool COptions::parseArguments(string_q& command) {
@@ -78,7 +78,7 @@ void COptions::Init(void) {
     nParamsRef = nParams;
     pOptions = this;
 
-    for (uint32_t i = 0 ; i < MAX_ADDRS ; i++) {
+    for (size_t i = 0 ; i < MAX_ADDRS ; i++) {
         addrs[i] = "";
     }
     nAddrs = 0;
